tplay第一个版本，希望大家用的开心。

+++ application/admin/controller/Attachment.php
@@ -53,14 +53,26 @@ class Attachment extends User
     {
     	if($this->request->isAjax()) {
     		$id = $this->request->has('id') ? $this->request->param('id', 0, 'intval') : 0;
-    		//$attachment = Db::name('attachment')->where('id',$id)->value('filepath');
-    		//$attachment = str_replace("\\","/",$attachment);
-    		if(false == Db::name('attachment')->where('id',$id)->delete()) {
-    			return $this->error('删除失败');
-    		} else {
-                addlog($id);//写入日志
-    			return $this->success('删除成功','admin/attachment/index');
-    		}
+    		$attachment = Db::name('attachment')->where('id',$id)->value('filepath');
+            if(file_exists($attachment)) {
+                if(unlink(ROOT_PATH . $attachment)) {
+                    if(false == Db::name('attachment')->where('id',$id)->delete()) {
+                        return $this->error('删除失败');
+                    } else {
+                        addlog($id);//写入日志
+                        return $this->success('删除成功','admin/attachment/index');
+                    }
+                } else {
+                    return $this->error('删除失败');
+                }
+            } else {
+                if(false == Db::name('attachment')->where('id',$id)->delete()) {
+                    return $this->error('删除失败');
+                } else {
+                    addlog($id);//写入日志
+                    return $this->success('删除成功','admin/attachment/index');
+                }
+            }
     	}
     }
